+++ src/db/db_mysql/mod.rs
@@ -160,7 +160,7 @@ impl MysqlConnection {
             let mut insert_query = vec![];
             let mut insert_params: Vec<Value> = vec![];
             for change in list {
-                insert_query.push("(?,?,?,?,?,?,?,?,?,?,?,'')");
+                insert_query.push("(?,?,?,?,?,?,?,?,?,?,?,'',UTC_TIMESTAMP())");
                 insert_params.push(change.name.clone().into());
                 insert_params.push(change.url.clone().into());
                 insert_params.push(change.homepage.clone().into());
@@ -174,7 +174,7 @@ impl MysqlConnection {
                 insert_params.push(change.stationuuid.clone().into());
                 list_ids.push(change.stationuuid.clone());
             }
-            let query = format!("INSERT INTO Station(Name,Url,Homepage,Favicon,Country,CountryCode,Subcountry,Language,Tags,ChangeUuid,StationUuid, UrlCache) 
+            let query = format!("INSERT INTO Station(Name,Url,Homepage,Favicon,Country,CountryCode,Subcountry,Language,Tags,ChangeUuid,StationUuid, UrlCache, Creation) 
                                     VALUES{}", insert_query.join(","));
             let mut stmt = transaction.prepare(query)?;
             stmt.execute(insert_params)?;
@@ -194,7 +194,7 @@ impl MysqlConnection {
 impl DbConnection for MysqlConnection {
     fn delete_old_checks(&mut self, hours: u32) -> Result<(), Box<dyn Error>> {
         let p = params!(hours);
-        let delete_old_checks_history_query = "DELETE FROM StationCheckHistory WHERE CheckTime < NOW() - INTERVAL :hours HOUR";
+        let delete_old_checks_history_query = "DELETE FROM StationCheckHistory WHERE CheckTime < UTC_TIMESTAMP() - INTERVAL :hours HOUR";
         let mut delete_old_checks_history_stmt = self.pool.prepare(delete_old_checks_history_query)?;
         delete_old_checks_history_stmt.execute(&p)?;
 
@@ -202,28 +202,28 @@ impl DbConnection for MysqlConnection {
     }
 
     fn delete_old_clicks(&mut self, hours: u32) -> Result<(), Box<dyn Error>> {
-        let delete_old_clicks_query = "DELETE FROM StationClick WHERE ClickTimestamp < NOW() - INTERVAL :hours HOUR";
+        let delete_old_clicks_query = "DELETE FROM StationClick WHERE ClickTimestamp < UTC_TIMESTAMP() - INTERVAL :hours HOUR";
         let mut delete_old_clicks_stmt = self.pool.prepare(delete_old_clicks_query)?;
         delete_old_clicks_stmt.execute(params!(hours))?;
         Ok(())
     }
 
     fn delete_never_working(&mut self, hours: u32) -> Result<(), Box<dyn Error>> {
-        let delete_never_working_query = "DELETE FROM Station WHERE LastCheckOkTime IS NULL AND Creation < NOW() - INTERVAL :hours HOUR";
+        let delete_never_working_query = "DELETE FROM Station WHERE LastCheckOkTime IS NULL AND Creation < UTC_TIMESTAMP() - INTERVAL :hours HOUR";
         let mut delete_never_working_stmt = self.pool.prepare(delete_never_working_query)?;
         delete_never_working_stmt.execute(params!(hours))?;
         Ok(())
     }
 
     fn delete_were_working(&mut self, hours: u32) -> Result<(), Box<dyn Error>> {
-        let delete_were_working_query = "DELETE FROM Station WHERE LastCheckOK=0 AND LastCheckOkTime IS NOT NULL AND LastCheckOkTime < NOW() - INTERVAL :hours HOUR";
+        let delete_were_working_query = "DELETE FROM Station WHERE LastCheckOK=0 AND LastCheckOkTime IS NOT NULL AND LastCheckOkTime < UTC_TIMESTAMP() - INTERVAL :hours HOUR";
         let mut delete_were_working_stmt = self.pool.prepare(delete_were_working_query)?;
         delete_were_working_stmt.execute(params!(hours))?;
         Ok(())
     }
 
     fn remove_unused_ip_infos_from_stationclicks(&mut self, hours: u32) -> Result<(), Box<dyn Error>> {
-        let query = "UPDATE StationClick SET IP=NULL WHERE InsertTime < NOW() - INTERVAL :hours HOUR";
+        let query = "UPDATE StationClick SET IP=NULL WHERE InsertTime < UTC_TIMESTAMP() - INTERVAL :hours HOUR";
         let mut stmt = self.pool.prepare(query)?;
         stmt.execute(params!(hours))?;
         Ok(())
@@ -234,8 +234,8 @@ impl DbConnection for MysqlConnection {
         clickcount=IFNULL((SELECT COUNT(*) FROM StationClick sc WHERE st.StationUuid=sc.StationUuid),0),
         ClickTrend=
         (
-            (select count(*) from StationClick sc1 where sc1.StationUuid=st.StationUuid AND ClickTimestamp>DATE_SUB(NOW(),INTERVAL 1 DAY) AND ClickTimestamp<=DATE_SUB(NOW(),INTERVAL 0 DAY)) - 
-            (select count(*) from StationClick sc2 where sc2.StationUuid=st.StationUuid AND ClickTimestamp>DATE_SUB(NOW(),INTERVAL 2 DAY) AND ClickTimestamp<=DATE_SUB(NOW(),INTERVAL 1 DAY))
+            (select count(*) from StationClick sc1 where sc1.StationUuid=st.StationUuid AND ClickTimestamp>DATE_SUB(UTC_TIMESTAMP(),INTERVAL 1 DAY) AND ClickTimestamp<=DATE_SUB(UTC_TIMESTAMP(),INTERVAL 0 DAY)) - 
+            (select count(*) from StationClick sc2 where sc2.StationUuid=st.StationUuid AND ClickTimestamp>DATE_SUB(UTC_TIMESTAMP(),INTERVAL 2 DAY) AND ClickTimestamp<=DATE_SUB(UTC_TIMESTAMP(),INTERVAL 1 DAY))
         ),
         ClickTimestamp=(SELECT Max(ClickTimestamp) FROM StationClick sc WHERE sc.StationUuid=st.StationUuid);";
         let mut stmt = self.pool.prepare(query)?;
@@ -265,19 +265,19 @@ impl DbConnection for MysqlConnection {
     }
 
     fn get_click_count_last_hour(&self) -> Result<u64, Box<dyn Error>> {
-        self.get_single_column_number(r#"SELECT COUNT(*) FROM StationClick WHERE TIMESTAMPDIFF(MINUTE,ClickTimestamp,now())<=60;"#)
+        self.get_single_column_number(r#"SELECT COUNT(*) FROM StationClick WHERE TIMESTAMPDIFF(MINUTE,ClickTimestamp,UTC_TIMESTAMP())<=60;"#)
     }
 
     fn get_click_count_last_day(&self) -> Result<u64, Box<dyn Error>> {
-        self.get_single_column_number(r#"SELECT COUNT(*) FROM StationClick WHERE TIMESTAMPDIFF(HOUR,ClickTimestamp,now())<=24;"#)
+        self.get_single_column_number(r#"SELECT COUNT(*) FROM StationClick WHERE TIMESTAMPDIFF(HOUR,ClickTimestamp,UTC_TIMESTAMP())<=24;"#)
     }
 
     fn get_station_count_todo(&self, hours: u32) -> Result<u64, Box<dyn Error>> {
-        self.get_single_column_number_params("SELECT COUNT(*) AS Items FROM Station WHERE LastLocalCheckTime IS NULL OR LastLocalCheckTime < NOW() - INTERVAL :hours HOUR", params!(hours))
+        self.get_single_column_number_params("SELECT COUNT(*) AS Items FROM Station WHERE LastLocalCheckTime IS NULL OR LastLocalCheckTime < UTC_TIMESTAMP() - INTERVAL :hours HOUR", params!(hours))
     }
 
     fn get_stations_to_check(&mut self, hours: u32, itemcount: u32) -> Result<Vec<StationItem>, Box<dyn Error>> {
-        let query = format!("SELECT {columns} FROM Station WHERE LastLocalCheckTime IS NULL OR LastLocalCheckTime < NOW() - INTERVAL {interval} HOUR ORDER BY RAND() LIMIT {limit}", columns = MysqlConnection::COLUMNS, interval = hours, limit = itemcount);
+        let query = format!("SELECT {columns} FROM Station WHERE LastLocalCheckTime IS NULL OR LastLocalCheckTime < UTC_TIMESTAMP() - INTERVAL {interval} HOUR ORDER BY RAND() LIMIT {limit}", columns = MysqlConnection::COLUMNS, interval = hours, limit = itemcount);
         let results = self.pool.prep_exec(query, ())?;
         self.get_list_from_query_result(results)
     }
@@ -292,15 +292,15 @@ impl DbConnection for MysqlConnection {
     }
 
     fn get_checks_todo_count(&self, hours: u32, source: &str) -> Result<u64, Box<dyn Error>> {
-        self.get_single_column_number_params("SELECT COUNT(*) AS Items FROM StationCheckHistory WHERE Source=:source AND CheckTime > NOW() - INTERVAL :hours HOUR",params!(hours, source))
+        self.get_single_column_number_params("SELECT COUNT(*) AS Items FROM StationCheckHistory WHERE Source=:source AND CheckTime > UTC_TIMESTAMP() - INTERVAL :hours HOUR",params!(hours, source))
     }
 
     fn get_deletable_never_working(&self, hours: u32) -> Result<u64, Box<dyn Error>> {
-        self.get_single_column_number_params("SELECT COUNT(*) AS Items FROM Station WHERE LastCheckOkTime IS NULL AND Creation < NOW() - INTERVAL :hours HOUR", params!(hours))
+        self.get_single_column_number_params("SELECT COUNT(*) AS Items FROM Station WHERE LastCheckOkTime IS NULL AND Creation < UTC_TIMESTAMP() - INTERVAL :hours HOUR", params!(hours))
     }
 
     fn get_deletable_were_working(&self, hours: u32) -> Result<u64, Box<dyn Error>> {
-        self.get_single_column_number_params("SELECT COUNT(*) AS Items FROM Station WHERE LastCheckOK=0 AND LastCheckOkTime IS NOT NULL AND LastCheckOkTime < NOW() - INTERVAL :hours HOUR", params!(hours))
+        self.get_single_column_number_params("SELECT COUNT(*) AS Items FROM Station WHERE LastCheckOK=0 AND LastCheckOkTime IS NOT NULL AND LastCheckOkTime < UTC_TIMESTAMP() - INTERVAL :hours HOUR", params!(hours))
     }
 
     fn get_stations_broken(&self, limit: u32) -> Result<Vec<StationItem>, Box<dyn Error>> {
@@ -559,7 +559,7 @@ impl DbConnection for MysqlConnection {
 
     fn get_changes(&self, stationuuid: Option<String>, changeuuid: Option<String>) -> Result<Vec<StationHistoryItem>, Box<dyn Error>> {
         let changeuuid_str = if changeuuid.is_some() {
-            " AND StationChangeID>=(SELECT StationChangeID FROM StationHistory WHERE ChangeUuid=:changeuuid) AND StationChangeID <= (SELECT MAX(StationChangeID) FROM StationHistory WHERE Creation <= NOW() - INTERVAL 60 SECOND) AND ChangeUuid<>:changeuuid"
+            " AND StationChangeID>=(SELECT StationChangeID FROM StationHistory WHERE ChangeUuid=:changeuuid) AND StationChangeID <= (SELECT MAX(StationChangeID) FROM StationHistory WHERE Creation <= UTC_TIMESTAMP() - INTERVAL 60 SECOND) AND ChangeUuid<>:changeuuid"
         } else {
             ""
         };
@@ -591,8 +591,8 @@ impl DbConnection for MysqlConnection {
         country: Option<String>, countrycode: Option<String>, state: Option<String>, language: Option<String>, tags: Option<String>) -> Result<String, Box<dyn Error>> {
         let mut transaction = self.pool.start_transaction(false, None, None)?;
 
-        let query = format!("INSERT INTO Station(Name,Url,Homepage,Favicon,Country,CountryCode,Subcountry,Language,Tags,ChangeUuid,StationUuid, UrlCache) 
-                        VALUES(:name, :url, :homepage, :favicon, :country, :countrycode, :state, :language, :tags, :changeuuid, :stationuuid, '')");
+        let query = format!("INSERT INTO Station(Name,Url,Homepage,Favicon,Country,CountryCode,Subcountry,Language,Tags,ChangeUuid,StationUuid, UrlCache,Creation) 
+                        VALUES(:name, :url, :homepage, :favicon, :country, :countrycode, :state, :language, :tags, :changeuuid, :stationuuid, '', UTC_TIMESTAMP())");
 
         let name = name.ok_or(DbError::AddStationError(String::from("name is empty")))?;
         let url = url.ok_or(DbError::AddStationError(String::from("url is empty")))?;
@@ -783,12 +783,30 @@ impl DbConnection for MysqlConnection {
                         continue;
                     }
                     // reuse checkuuid
-                    insert_station_check_query.push("(?,?,?,?,?,?,?,NOW(),?,?,?,?,?,?,?,?,?,?)");
-                    insert_station_check_params.push(checkuuid.into());
+                    match &item.timestamp {
+                        Some(timestamp) => {
+                            insert_station_check_query.push("(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,UTC_TIMESTAMP())");
+                            insert_station_check_params.push(checkuuid.into());
+                            insert_station_check_params.push(timestamp.into());
+                        }
+                        None => {
+                            insert_station_check_query.push("(?,UTC_TIMESTAMP(),?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,UTC_TIMESTAMP())");
+                            insert_station_check_params.push(checkuuid.into());
+                        }
+                    }
                 },
                 None => {
                     // generate new checkuuid
-                    insert_station_check_query.push("(UUID(),?,?,?,?,?,?,NOW(),?,?,?,?,?,?,?,?,?,?)");
+                    match &item.timestamp {
+                        Some(timestamp) => {
+                            insert_station_check_query.push("(UUID(),?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,UTC_TIMESTAMP())");
+                            insert_station_check_params.push(timestamp.into());
+                        }
+                        None => {
+                            insert_station_check_query.push("(UUID(),UTC_TIMESTAMP(),?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,UTC_TIMESTAMP())");
+                        }
+                    }
+                    
                 }
             }
             
@@ -818,8 +836,8 @@ impl DbConnection for MysqlConnection {
         // insert into history table
         if insert_station_check_query.len() > 0 {
             let insert_station_check_params_str = insert_station_check_query.join(",");
-            let query_insert_station_check_history = format!("INSERT INTO StationCheckHistory(CheckUuid,StationUuid,Source,Codec,Bitrate,Hls,CheckOK,CheckTime,UrlCache,
-                MetainfoOverridesDatabase,Public,Name,Description,Tags,CountryCode,Homepage,Favicon,Loadbalancer) VALUES{}", insert_station_check_params_str);
+            let query_insert_station_check_history = format!("INSERT INTO StationCheckHistory(CheckUuid,CheckTime,StationUuid,Source,Codec,Bitrate,Hls,CheckOK,UrlCache,
+                MetainfoOverridesDatabase,Public,Name,Description,Tags,CountryCode,Homepage,Favicon,Loadbalancer,InsertTime) VALUES{}", insert_station_check_params_str);
             let mut stmt_insert_station_check_history = transaction.prepare(query_insert_station_check_history)?;
             stmt_insert_station_check_history.execute(&insert_station_check_params)?;
         }
@@ -897,11 +915,11 @@ impl DbConnection for MysqlConnection {
                         }
                         query.push("LastCheckOk=:vote");
                         if local {
-                            query.push("LastLocalCheckTime=NOW()");
+                            query.push("LastLocalCheckTime=UTC_TIMESTAMP()");
                         }
 
                         if item.check_ok {
-                            let query_update_ok = format!("UPDATE Station SET LastCheckOkTime=NOW(),LastCheckTime=NOW(),Codec=:codec,Bitrate=:bitrate,Hls=:hls,UrlCache=:urlcache,{} WHERE StationUuid=:stationuuid", query.join(","));
+                            let query_update_ok = format!("UPDATE Station SET LastCheckOkTime=UTC_TIMESTAMP(),LastCheckTime=UTC_TIMESTAMP(),Codec=:codec,Bitrate=:bitrate,Hls=:hls,UrlCache=:urlcache,{} WHERE StationUuid=:stationuuid", query.join(","));
                             let mut stmt_update_ok = transaction.prepare(query_update_ok)?;
                             stmt_update_ok.execute(params)?;
                         }
@@ -914,14 +932,14 @@ impl DbConnection for MysqlConnection {
                     if item.check_ok {
                         params.push((String::from("urlcache"), item.url.clone().into(),));
 
-                        let query_update_ok = format!("UPDATE Station SET {lastlocalchecktime}LastCheckOkTime=NOW(),LastCheckTime=NOW(),Codec=:codec,Bitrate=:bitrate,Hls=:hls,UrlCache=:urlcache,LastCheckOk=:vote WHERE StationUuid=:stationuuid",
-                            lastlocalchecktime = if local {"LastLocalCheckTime=NOW(),"} else {""},
+                        let query_update_ok = format!("UPDATE Station SET {lastlocalchecktime}LastCheckOkTime=UTC_TIMESTAMP(),LastCheckTime=UTC_TIMESTAMP(),Codec=:codec,Bitrate=:bitrate,Hls=:hls,UrlCache=:urlcache,LastCheckOk=:vote WHERE StationUuid=:stationuuid",
+                            lastlocalchecktime = if local {"LastLocalCheckTime=UTC_TIMESTAMP(),"} else {""},
                         );
                         let mut stmt_update_ok = transaction.prepare(query_update_ok)?;
                         stmt_update_ok.execute(params)?;
                     }else{
-                        let query_update_check_ok = format!("UPDATE Station st SET {lastlocalchecktime}LastCheckTime=NOW() WHERE StationUuid=:stationuuid",
-                            lastlocalchecktime = if local {"LastLocalCheckTime=NOW(),"} else {""},
+                        let query_update_check_ok = format!("UPDATE Station st SET {lastlocalchecktime}LastCheckTime=UTC_TIMESTAMP() WHERE StationUuid=:stationuuid",
+                            lastlocalchecktime = if local {"LastLocalCheckTime=UTC_TIMESTAMP(),"} else {""},
                         );
                         let mut stmt_update_check_ok = transaction.prepare(query_update_check_ok)?;
                         stmt_update_check_ok.execute(params)?;
@@ -964,12 +982,12 @@ impl DbConnection for MysqlConnection {
                 insert_click_params.push(item.stationuuid.clone().into());
                 insert_click_params.push(item.clicktimestamp.clone().into());
 
-                insert_click_query.push("(?,?,?)");
+                insert_click_query.push("(?,?,?,UTC_TIMESTAMP())");
             }
         }
 
         if insert_click_query.len() > 0 {
-            let query = format!("INSERT INTO StationClick(ClickUuid, StationUuid, ClickTimestamp) VALUES{}", insert_click_query.join(","));
+            let query = format!("INSERT INTO StationClick(ClickUuid, StationUuid, ClickTimestamp, InsertTime) VALUES{}", insert_click_query.join(","));
             let mut stmt_insert = transaction.prepare(query)?;
             stmt_insert.execute(insert_click_params)?;
         }
@@ -983,7 +1001,7 @@ impl DbConnection for MysqlConnection {
         let table_name = if include_history { "StationCheckHistory" } else { "StationCheck" };
         let where_seconds = if seconds > 0 {
             format!(
-                "AND TIMESTAMPDIFF(SECOND,CheckTime,now())<{seconds}",
+                "AND TIMESTAMPDIFF(SECOND,CheckTime,UTC_TIMESTAMP())<{seconds}",
                 seconds = seconds
             )
         } else {
@@ -995,7 +1013,7 @@ impl DbConnection for MysqlConnection {
                 let mut query_params = params!{"stationuuid" => stationuuid};
                 let where_checkuuid_str = if checkuuid.is_some() {
                     query_params.push((String::from("checkuuid"), checkuuid.unwrap().into(),));
-                    format!(" AND CheckID>=(SELECT CheckID FROM {table_name} WHERE CheckUuid=:checkuuid) AND CheckID <= (SELECT MAX(CheckID) FROM {table_name} WHERE InsertTime <= NOW() - INTERVAL 60 SECOND) AND CheckUuid<>:checkuuid", table_name = table_name)
+                    format!(" AND CheckID>=(SELECT CheckID FROM {table_name} WHERE CheckUuid=:checkuuid) AND CheckID <= (SELECT MAX(CheckID) FROM {table_name} WHERE InsertTime <= UTC_TIMESTAMP() - INTERVAL 60 SECOND) AND CheckUuid<>:checkuuid", table_name = table_name)
                 } else {
                     String::from("")
                 };
@@ -1009,7 +1027,7 @@ impl DbConnection for MysqlConnection {
                 let where_checkuuid_str = match checkuuid {
                     Some(checkuuid) => {
                         query_params.push((String::from("checkuuid"), checkuuid.into(),));
-                        format!(" AND CheckID>=(SELECT CheckID FROM {table_name} WHERE CheckUuid=:checkuuid) AND CheckID <= (SELECT MAX(CheckID) FROM {table_name} WHERE InsertTime <= NOW() - INTERVAL 60 SECOND) AND CheckUuid<>:checkuuid", table_name = table_name)
+                        format!(" AND CheckID>=(SELECT CheckID FROM {table_name} WHERE CheckUuid=:checkuuid) AND CheckID <= (SELECT MAX(CheckID) FROM {table_name} WHERE InsertTime <= UTC_TIMESTAMP() - INTERVAL 60 SECOND) AND CheckUuid<>:checkuuid", table_name = table_name)
                     },
                     None => String::from("")
                 };
@@ -1026,7 +1044,7 @@ impl DbConnection for MysqlConnection {
     fn get_clicks(&self, stationuuid: Option<String>, clickuuid: Option<String>, seconds: u32) -> Result<Vec<StationClickItem>, Box<dyn Error>> {
         let where_seconds = if seconds > 0 {
             format!(
-                "AND TIMESTAMPDIFF(SECOND,ClickTimestamp,now())<{seconds}",
+                "AND TIMESTAMPDIFF(SECOND,ClickTimestamp,UTC_TIMESTAMP())<{seconds}",
                 seconds = seconds
             )
         } else {
@@ -1039,7 +1057,7 @@ impl DbConnection for MysqlConnection {
                 let where_clickuuid_str = match clickuuid {
                     Some(clickuuid) => {
                         query_params.push((String::from("clickuuid"), clickuuid.into(),));
-                        " AND ClickID>=(SELECT ClickID FROM StationClick WHERE ClickUuid=:clickuuid) AND ClickID <= (SELECT MAX(ClickID) FROM StationClick WHERE InsertTime <= NOW() - INTERVAL 60 SECOND) AND ClickUuid<>:clickuuid"
+                        " AND ClickID>=(SELECT ClickID FROM StationClick WHERE ClickUuid=:clickuuid) AND ClickID <= (SELECT MAX(ClickID) FROM StationClick WHERE InsertTime <= UTC_TIMESTAMP() - INTERVAL 60 SECOND) AND ClickUuid<>:clickuuid"
                     },
                     None => ""
                 };
@@ -1052,7 +1070,7 @@ impl DbConnection for MysqlConnection {
                 let mut query_params = params!{"one" => 1};
                 let where_clickuuid_str = if clickuuid.is_some() {
                     query_params.push((String::from("clickuuid"), clickuuid.unwrap().into(),));
-                    " AND ClickID>=(SELECT ClickID FROM StationClick WHERE ClickUuid=:clickuuid) AND ClickID <= (SELECT MAX(ClickID) FROM StationClick WHERE InsertTime <= NOW() - INTERVAL 60 SECOND) AND ClickUuid<>:clickuuid"
+                    " AND ClickID>=(SELECT ClickID FROM StationClick WHERE ClickUuid=:clickuuid) AND ClickID <= (SELECT MAX(ClickID) FROM StationClick WHERE InsertTime <= UTC_TIMESTAMP() - INTERVAL 60 SECOND) AND ClickUuid<>:clickuuid"
                 } else {
                     ""
                 };
@@ -1286,12 +1304,8 @@ impl DbConnection for MysqlConnection {
                 let _result_1_delete = self.pool.prep_exec(query_1_delete, ())?;
 
                 // was there a vote from the ip in the last 1 day?
-                let query_2_vote_check = format!(
-                    r#"SELECT StationID FROM IPVoteCheck WHERE StationID={id} AND IP="{ip}""#,
-                    id = station.id,
-                    ip = ip
-                );
-                let result_2_vote_check = self.pool.prep_exec(query_2_vote_check, ())?;
+                let query_2_vote_check = "SELECT StationID FROM IPVoteCheck WHERE StationID=:id AND IP=:ip";
+                let result_2_vote_check = self.pool.prep_exec(query_2_vote_check, params!(ip, "id" => station.id))?;
                 for resultsingle in result_2_vote_check {
                     for _ in resultsingle {
                         // do not allow vote
@@ -1300,23 +1314,16 @@ impl DbConnection for MysqlConnection {
                 }
 
                 // add vote entry
-                let query_3_insert_votecheck = format!(
-                    r#"INSERT INTO IPVoteCheck(IP,StationID) VALUES("{ip}",{id})"#,
-                    id = station.id,
-                    ip = ip
-                );
+                let query_3_insert_votecheck = "INSERT INTO IPVoteCheck(IP,StationID,VoteTimestamp) VALUES(:ip,:id,UTC_TIMESTAMP())";
                 let result_3_insert_votecheck =
-                    self.pool.prep_exec(query_3_insert_votecheck, ())?;
+                    self.pool.prep_exec(query_3_insert_votecheck, params!(ip,"id" => station.id))?;
                 if result_3_insert_votecheck.affected_rows() == 0 {
                     return Err(Box::new(DbError::VoteError("could not insert vote check".to_string())));
                 }
 
                 // vote for station
-                let query_4_update_votes = format!(
-                    "UPDATE Station SET Votes=Votes+1 WHERE StationID={id}",
-                    id = station.id
-                );
-                let result_4_update_votes = self.pool.prep_exec(query_4_update_votes, ())?;
+                let query_4_update_votes = "UPDATE Station SET Votes=Votes+1 WHERE StationID=:id";
+                let result_4_update_votes = self.pool.prep_exec(query_4_update_votes, params!("id" => station.id))?;
                 if result_4_update_votes.affected_rows() == 1 {
                     Ok("voted for station successfully".to_string())
                 } else {
@@ -1335,10 +1342,10 @@ impl DbConnection for MysqlConnection {
             return Ok(false);
         }
 
-        let query2 = "INSERT INTO StationClick(IP,StationUuid,ClickUuid) VALUES(:ip,:stationuuid,UUID())";
+        let query2 = "INSERT INTO StationClick(IP,StationUuid,ClickUuid,ClickTimestamp,InsertTime) VALUES(:ip,:stationuuid,UUID(),UTC_TIMESTAMP(),UTC_TIMESTAMP())";
         let result2 = self.pool.prep_exec(query2, params!{"stationuuid" => &station.stationuuid, "ip" => ip})?;
 
-        let query3 = "UPDATE Station SET ClickTimestamp=NOW() WHERE StationUuid=:stationuuid";
+        let query3 = "UPDATE Station SET ClickTimestamp=UTC_TIMESTAMP() WHERE StationUuid=:stationuuid";
         let result3 = self.pool.prep_exec(query3, params!{"stationuuid" => &station.stationuuid})?;
 
         if result2.affected_rows() == 1 && result3.affected_rows() == 1 {